refactor :: 다운로드 파일 열기 함수 추가
